Ajout de messages d'erreurs visibles par le joueur

+++ app/Pioche.php
@@ -44,27 +44,32 @@ class Pioche extends Model
 
     //Au début faire tirerCarte($id,-1)
     public static function tirerCarte($idpartie,$idjoueur) {
-      $carte=Pioche::where('idpartie', $idpartie)->where('position', 1)->get()->first();
-      $typecarte=$carte->typecarte;
-
-      //Suppression de la carte de la Pioche
-      $carte->delete();
-
-      //Décalage de toutes les positions
       $size=Pioche::where('idpartie',$idpartie)->get()->count();
-      for($i=0;$i<$size;$i++){
-        $carte=Pioche::where('idpartie',$idpartie)->get()[$i];
-        $carte->position=$carte->position-1;
-        $carte->save();
+
+      if ($size > 0) {
+        $carte=Pioche::where('idpartie', $idpartie)->where('position', 1)->get()->first();
+        $typecarte=$carte->typecarte;
+
+        //Suppression de la carte de la Pioche
+        $carte->delete();
+
+        //Décalage de toutes les positions
+        $size=Pioche::where('idpartie',$idpartie)->get()->count();
+        for($i=0;$i<$size;$i++){
+          $carte=Pioche::where('idpartie',$idpartie)->get()[$i];
+          $carte->position=$carte->position-1;
+          $carte->save();
+        }
+
+        //Si l'idjoueur n'est pas spécifié la carte va directement dans la défausse
+        if ($idjoueur == -1) {
+          Defausse::defausserCarte($idpartie,$typecarte,-1);
+        } else {
+          //On ajoute la carte à la main du joueur
+          Main::tirerCarte($idpartie,$idjoueur,$typecarte);
+        }
       }
 
-      //Si l'idjoueur n'est pas spécifié la carte va directement dans la défausse
-      if ($idjoueur == -1) {
-        Defausse::defausserCarte($idpartie,$typecarte,-1);
-      } else {
-        //On ajoute la carte à la main du joueur
-        Main::tirerCarte($idpartie,$idjoueur,$typecarte);
-      }
     }
 
     public static function nbPioche($idpartie) {
